requirement.txt
Django
django-cors-headers
django-filter
djangorestframework
djangorestframework-jwt
redis